Move pylxd.model to pylxd.models._model
