+++ pylxd/models/_model.py
