+++ pylxd/tests/test_model.py
@@ -11,7 +11,7 @@
 #    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 #    License for the specific language governing permissions and limitations
 #    under the License.
-from pylxd import model
+from pylxd.models import _model as model
 from pylxd.tests import testing
